VNNewsVoice_CrawlingService/pyproject.toml
[project]
name = "vnnewsvoice-crawlingservice"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aio-pika==9.4.3",
    "aiohttp==3.11.18",
    "beautifulsoup4==4.13.4",
    "fastapi==0.115.14",
    "feedparser==6.0.11",
    "lxml==6.0.0",
    "orjson==3.10.18",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "ruff>=0.15.6",
    "tenacity==8.5.0",
    "uvicorn[standard]==0.35.0",
    "prometheus-fastapi-instrumentator>=0.6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
